Fix CloudFront invalidation paths in development workflow
The CloudFront invalidation was using incorrect paths that included the
origin path prefix. Since the origin is configured with '/development'
path, the invalidation paths should be '/images/*' not '/development/images/*'.

This fixes the issue where new builds weren't properly invalidating the
CloudFront cache, causing the website to serve stale content.